[#16][Fix] compute valid queues and queue handles
When creating queues, we need information about which queues were created and at which queue family index. This information is lost when our previous getQueueCreateInfos function terminated. The actual idea is to save this information for each queue type and compute valid handles for them.

The code might look uglier than before... but maybe we can improve this. :D We also need to clean up the unnecessary print outs later on.