[#42] add vertical movement via E and Q keys to Camera
Before, the camera object was just able to move along the xz plane. For vertical movement, you have to pan view and then use forward or backward movement to move up or down. This might be not so handy for the user.
Now, you can use E to move up and use Q to move down (both along default world up vector).