[#42][Fix] fix up vector, camera coordinate system, depth
Vulkan uses a left handed coordinate system. Thus, we need to define GLM_FORCE_LEFT_HANDED and adjust left and right movement of the camera.

We also need to define GLM_DEPTH_ZERO_TO_ONE for the depth values which range between 0 and 1 in Vulkan.

Although the up vector is pointing upwards again, it seems that the rendered scene is not flipped along y axis... is this correct?