[#92] add RTX Pipeline creation function, make dispatcher member accessible for RTXModule (see also text below)
Problem:
The vk::Pipeline object cannot be used for scene.recordDrawcalls! Here, we need a vkcv::PipelineHandle. Maybe we need to move all functionalities from here to the PipelineManager - similar to PipelineManager::createComputePipeline and PipelineManager::createGraphicsPipeline. As far as I know, there are also some updates for the PipelineManager in develop. We should merge and see how to deal with the pipeline.