[#92] fixed acceleration structure creation.
I revisited the acceleration structure creation and tried different things and now it finally works (not perfect, but it works). The main issue was (i think) the data being given as uint8_t. I changed it to float for vertices and uint32_t for indices and now it works. Additionally i rewrote the way we submit commands to the commandbuffer.